requirements-supabase.txt
psycopg2-binary==2.9.9
supabase==2.0.1
python-dotenv==1.0.0
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Werkzeug==2.3.7
Flask-Mail==0.9.1
requests==2.31.0
groq==0.4.1
gunicorn==21.2.0
